backend/requirements.txt
fastapi
uvicorn
msal
requests
openai
pydantic
python-dotenv